Did some plumbing changes in our model
