+++ app/src/main/java/com/cmput301f18t20/medicalphotorecord/Provider.java
@@ -17,23 +17,20 @@ import java.util.NoSuchElementException;
 
 import Exceptions.UserIDMustBeAtLeastEightCharactersException;
 
-public class Provider extends User implements Refreshable {
+public class Provider extends User {
     protected ArrayList<Patient> patients = new ArrayList<>();
 
     public Provider(String userId)
             throws UserIDMustBeAtLeastEightCharactersException {
         super(userId);
-        this.refresh();
     }
 
     public Provider(String userId, String email, String phoneNumber)
             throws UserIDMustBeAtLeastEightCharactersException {
         super(userId, email, phoneNumber);
-        this.refresh();
     }
 
     public Patient getPatient(int PatientIndex) {
-        this.refresh();
         if (PatientIndex > this.patients.size()) {
             throw new NoSuchElementException("Patient not found");
         }
@@ -41,7 +38,6 @@ public class Provider extends User implements Refreshable {
     }
 
     public Patient getPatient(String userID) {
-        this.refresh();
         for (Patient patient : this.patients) {
             if (patient.getUserID().equals(userID)) {
                 return patient;
@@ -51,24 +47,16 @@ public class Provider extends User implements Refreshable {
     }
 
     public ArrayList<Patient> getPatients() {
-        this.refresh();
         return patients;
     }
 
     public void assignPatient(Patient patient) {
         //TODO should we have a check here to make sure the same patient isn't added twice?
-        //FIXME should refresh be called before or after? Does it fetch from or write to the database?
         patients.add(patient);
     }
 
-    public void assignPatient(String userID) {
-        //Some sort of database lookup?
-    }
-
     public void unAssignPatient(Patient patient) {
-        patient.removeProvider(this);
         this.patients.remove(patient);
-        //TODO         this.refresh() before or after??;
     }
 
     public void unAssignPatient(String userID) {
@@ -76,16 +64,6 @@ public class Provider extends User implements Refreshable {
     }
 
     public void unAssignPatient(int PatientIndex) {
-        //lock database
-        //pull database changes
         this.patients.remove(PatientIndex);
-        //write to database
-        //TODO         this.refresh() before or after??;
-
-    }
-
-    @Override
-    public void refresh() {
-        //TODO
     }
 }
