--- app/src/main/java/com/cmput301f18t20/medicalphotorecord/QueryTool.java
@@ -1,24 +1,0 @@
-/*
- * Class name: QueryTool
- *
- * Version: Version 1.0
- *
- * Developed by members of CMPUT301F18T20 on Date: 11/11/18 6:21 PM
- *
- * Last Modified: 08/11/18 7:47 AM
- *
- * Copyright (c) 2018, CMPUT301F18T20, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
- */
-
-package com.cmput301f18t20.medicalphotorecord;
-
-public class QueryTool {
-    public void queryRecords(String[] UserIDs, String keywords, Filter filter) {}
-    public void queryRecords(String UserID, String keywords, Filter filter) {}
-
-    public void queryProblems(String[] UserIDs, String keywords, Filter filter) {}
-    public void queryProblems(String UserID, String keywords, Filter filter) {}
-
-    public void queryProblemsAndRecords(String[] UserIDs, String keywords, Filter filter) {}
-    public void queryProblemsAndRecords(String UserID, String keywords, Filter filter) {}
-}
