+++ app/src/test/java/com/cmput301f18t20/medicalphotorecord/PatientTests.java
@@ -14,10 +14,6 @@ import static org.junit.Assert.fail;
 public class PatientTests {
     protected static String CorrectUserID1 = "abcdefgh";
     protected static String CorrectUserID2 = "stuvwxyz";
-    protected static String CorrectUserID3 = "acebgfIII";
-
-
-
 
     @Test(expected = Test.None.class /* no exception expected */)
     public void testGetProblem()
